requirements.txt
django
django-crispy-forms
django-extensions
django-heroku
django-markdownx
django-rest-framework
django-simple-menu
djangorestframework
djangorestframework-stubs
gunicorn
plotly
psycopg2-binary
uuid
setuptools
